mutable self wip
