+++ ast/src/common/keyword_or_identifier.rs
@@ -15,27 +15,29 @@
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    ast::{span_into_string, Rule},
-    SpanDef,
+    ast::Rule,
+    common::{Identifier, SelfKeyword},
+    functions::InputKeyword,
 };
 
-use pest::Span;
 use pest_ast::FromPest;
 use serde::Serialize;
 use std::fmt;
 
 #[derive(Clone, Debug, FromPest, PartialEq, Serialize)]
-#[pest_ast(rule(Rule::self_keyword))]
-pub struct SelfKeyword<'ast> {
-    #[pest_ast(outer(with(span_into_string)))]
-    pub keyword: String,
-    #[pest_ast(outer())]
-    #[serde(with = "SpanDef")]
-    pub span: Span<'ast>,
+#[pest_ast(rule(Rule::keyword_or_identifier))]
+pub enum KeywordOrIdentifier<'ast> {
+    SelfKeyword(SelfKeyword<'ast>),
+    Input(InputKeyword<'ast>),
+    Identifier(Identifier<'ast>),
 }
 
-impl<'ast> fmt::Display for SelfKeyword<'ast> {
+impl<'ast> fmt::Display for KeywordOrIdentifier<'ast> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.keyword)
+        match self {
+            KeywordOrIdentifier::SelfKeyword(self_keyword) => write!(f, "{}", self_keyword),
+            KeywordOrIdentifier::Input(input_keyword) => write!(f, "{}", input_keyword),
+            KeywordOrIdentifier::Identifier(identifier) => write!(f, "{}", identifier),
+        }
     }
 }
