+++ compiler/src/statement/assign/assign.rs
@@ -79,8 +79,8 @@ impl<F: Field + PrimeField, G: GroupType<F>> ConstrainedProgram<F, G> {
                 span,
             ),
             Assignee::Tuple(_tuple, index) => self.assign_tuple(cs, indicator, variable_name, index, new_value, span),
-            Assignee::CircuitField(_assignee, object_name) => {
-                self.mutute_circuit_variable(cs, indicator, variable_name, object_name, new_value, span)
+            Assignee::CircuitField(_assignee, circuit_variable) => {
+                self.mutute_circuit_variable(cs, indicator, variable_name, circuit_variable, new_value, span)
             }
         }
     }
